fix convert to procurement
For products with route=Buy, in a warehouse with 2 step receipt, we can't send the stock location (usually
specified on the orderpoint) to the procurement. Otherwise, the input -> stock procurement rules is selected,
and a stock move is created for the internal transfer from input to stock.

In this case, one of two problems arises:
    (1) if procurement group is not specified, then the stock move is attached to any open stock picking that
        has no procurement group
    (2) if a procurement group is specified, then a separate picking is created for each procurement item.

When creating the procurement, we give a specific procurement rule, rather than specifying a location and
letting the procurement find the rule.

Our search for an appropriate procurement rule uses the product's route and the orderpoint's warehouse. If we
can't find a procurement rule, or if the product has multiple routes, we will notify the user and we will not
create the procurement.  In this case, the user will either create the procurement manually or fix the product's
routes.

This way, the stock move and picking for the transfer is created by, and grouped with, the purchase order.